Created a StageFactory class to clean the code base.
Add a falsetto detector, tuned it for better accuracy in the python script.

Prevented the application from intentionally crashing when the audio recording goes missing, set an alert to indicate the issue to the user.

Refactored the code base so that the stages used the StageFactory.

Add invalidation enumerator to detect if the audio file was invalid or if the python script returned a system code 1.

Added ui element to the loading screen to output where the current process is when loading.